+++ flutter_lmdb2/macos/Classes/FlutterLmdb2Plugin.swift
@@ -1,0 +1,21 @@
+import FlutterMacOS
+import Foundation
+
+public class FlutterLmdb2Plugin: NSObject, FlutterPlugin {
+    public static func register(with registrar: FlutterPluginRegistrar) {
+        let channel = FlutterMethodChannel(
+            name: "flutter_lmdb2",
+            binaryMessenger: registrar.messenger)
+        let instance = FlutterLmdb2Plugin()
+        registrar.addMethodCallDelegate(instance, channel: channel)
+    }
+
+    public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+        switch call.method {
+        case "getPlatformVersion":
+            result("macOS " + ProcessInfo.processInfo.operatingSystemVersionString)
+        default:
+            result(FlutterMethodNotImplemented)
+        }
+    }
+}